Reorganized files into separate folders
