+++ jobs/Backend/Task/src/ExchangeRateUpdater/GlobalUsings.cs
@@ -1,0 +1,6 @@
+﻿global using ExchangeRateUpdater.Contracts;
+global using ExchangeRateUpdater.Models;
+global using ExchangeRateUpdater.Services;
+global using System.Collections.Generic;
+global using System.Linq;
+global using System;
