+++ jobs/Backend/Task/src/ExchangeRateUpdater/Models/ExchangeRate.cs
@@ -1,4 +1,4 @@
-﻿namespace ExchangeRateUpdater;
+﻿namespace ExchangeRateUpdater.Models;
 
 public class ExchangeRate
 {
