+++ jobs/Backend/Task/src/ExchangeRateUpdater/Program.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using ExchangeRateUpdater.Models;
+using ExchangeRateUpdater.Services;
 
 namespace ExchangeRateUpdater;
